refactor(platform): extract agent delete dialog to fix eslint complexity
ZeroSettingsTab exceeded the complexity limit (22 > 20) after adding the
two-column split delete dialog. Move the danger-zone card and delete
reconcile flow into a dedicated AgentDeleteDialog component, dropping the
parent well under the limit without any eslint-disable.
